+++ src/app/app-routing.module.ts
@@ -4,6 +4,7 @@ import { DashboardComponent } from './dashboard/dashboard.component';
 import { HeroesComponent } from './heroes/heroes.component';
 import { HeroDetailComponent } from './hero-detail/hero-detail.component';
 import { TenderNamesResolver } from './resolvers/tender-names/tender-names.resolver';
+import { RedirectToDashboardResolver } from './resolvers/redirect-to-dashboard/redirect-to-dashboard.resolver';
 
 @NgModule({
   imports: [
@@ -21,6 +22,14 @@ import { TenderNamesResolver } from './resolvers/tender-names/tender-names.resol
         component: HeroDetailComponent,
         resolve: {
           tenderNames: TenderNamesResolver
+        }
+      },
+      {
+        status: [999],
+        type: '*',
+        component: DashboardComponent,
+        resolve: {
+          toBase: RedirectToDashboardResolver
         }
       }
     ])